+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Version 1.0.2, 2019/02/04
 
   - miner-coexpr tool added
   - miner-clusters tool added
+  - miner-enrichment tool added
 
 Version 1.0.1, 2019/01/22
 -------------------------